% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curtain.R
\name{curtain}
\alias{curtain}
\title{Create curtain widget}
\usage{
curtain(
  colors,
  width = 512,
  height = 256,
  margin = 0,
  background = NULL,
  wave_width = 8,
  wave_amplitude = 0.125,
  wave_exponent = 2,
  n_color_interp = 512,
  elementId = NULL
)
}
\arguments{
\item{colors}{\code{character} vector of CSS colors, e.g. hex-codes.}

\item{width}{\code{numeric}   width of canvas (pixels).}

\item{height}{\code{numeric}   height of canvas (pixels).}

\item{margin}{\code{numeric}   margin around canvas (pixels).}

\item{background}{\code{character} CSS color for canvas background.}

\item{wave_width}{\code{numeric}   width of sine wave, i.e. wavelength
(pixels).}

\item{wave_amplitude}{\code{numeric}   amplitude of sine wave (non-dimensional).}

\item{wave_exponent}{\code{numeric}   exponent for vertical dampening of
sine wave (non-dimensional):
use 0 for no dampening,
use 1 for linear dampening.}

\item{n_color_interp}{\code{numeric}   number interpolated colors on canvas.}

\item{elementId}{\code{character} ID for enclosing element.}
}
\value{
Object with S3 classes \code{splattr_curtain}, \code{htmlwidget}.
}
\description{
Acknowledge \code{\link[pals:pal.sineramp]{pals::pal.sineramp()}}, Peter Kovesi.
}
\section{Widget Sizing}{


The canvas width can be decreased according to the enclosing container, but will not exceed \code{width}.
}

