% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{df_hcl}
\alias{df_hcl}
\title{Create HCL data frame}
\usage{
df_hcl(colors)
}
\arguments{
\item{colors}{\code{character} vector of R colors. Can be hex codes, etc.}
}
\value{
\code{data.frame} with columns: \code{color}, \code{hue}, \code{chroma} \code{luminance}.
}
\description{
Needs colorspace package.
}
\examples{
  df_hcl("red")
  df_hcl("#FF0000")
  df_hcl(c("red", "green"))
}
