% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cvd.R
\name{input_cvd}
\alias{input_cvd}
\title{Create CVD widget}
\usage{
input_cvd(
  choice = c("none", "deutan", "protan", "tritan"),
  severity = 0,
  label = "Color-Vision Deficiency",
  elementId = NULL
)
}
\arguments{
\item{choice}{\code{character} initialization: indicates which form of CVD to simulate.}

\item{severity}{\code{numeric} initialization: indicates the severity of CVD to simulate.}

\item{label}{\code{character} label for the widget.}

\item{elementId}{\code{character} ID for the enclosing element.}
}
\value{
htmlwidget
}
\description{
This is a set of inputs that offers a transformation matrix
for color-vision deficiency.
}
\examples{
  # input_cvd("cvd-control")
}
