% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramp.R
\name{ramp}
\alias{ramp}
\title{Create ramp widget}
\usage{
ramp(
  colors,
  n_color_widget = 512,
  height = 40,
  margin = 0,
  background = NULL,
  elementId = NULL
)
}
\arguments{
\item{colors}{\code{character} vector of CSS colors. Can be hex codes, etc.}

\item{n_color_widget}{\code{numeric} number of colors used in the widget.}

\item{height}{\code{numeric} height of canvas (pixels).}

\item{margin}{\code{numeric} margin around canvas (pixels).}

\item{background}{\code{character} CSS color for the canvas background.}

\item{elementId}{\code{character} ID for the enclosing element.}
}
\value{
htmlwidget
}
\description{
Useful for showing continuous palettes
}
\examples{
  ramp(terrain.colors(100))
}
