% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramp.R
\name{ramp}
\alias{ramp}
\title{Create ramp widget}
\usage{
ramp(colors, n = length(colors), height = 40, elementId = NULL)
}
\arguments{
\item{colors}{\code{character} vector of CSS colors. Can be hex codes, etc.}

\item{n}{\code{numeric} number of colors to show in the ramp.
Uses \code{d3.interpolateRgbBasis()} to interpolate.}

\item{height}{\code{numeric} height of widget (pixels).}

\item{elementId}{\code{character} ID for the enclosing element.}
}
\value{
htmlwidget
}
\description{
Useful for showing continuous palettes
}
\examples{
  ramp(terrain.colors(100))
}
