% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swatch.R
\name{swatch}
\alias{swatch}
\title{Create swatch widget}
\usage{
swatch(colors, height = 40, dx = 10, elementId = NULL)
}
\arguments{
\item{colors}{\code{character} vector of CSS colors. Can be hex codes, etc.}

\item{height}{\code{numeric} height of widget (pixels).}

\item{dx}{\code{numeric} spacing between color-swatches (pixels).}

\item{elementId}{\code{character} ID for the enclosing element.}
}
\value{
htmlwidget
}
\description{
Useful for showing discrete palettes
}
\examples{
  swatch(terrain.colors(10))
}
